requirements.txt
pytest==7.4.0
pytest-bdd==6.1.1
pytest-html==3.2.0
paho-mqtt==1.6.1
